requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
psycopg2-binary==2.9.9

# Optional but recommended
django-extensions==3.2.3
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
